Update hoist tests to use in-repo run.sh
This converts the hoist repo to use the .bonnyci/run.sh entry point.
It also updates the bonnyci.yaml layout to allow passing the node
parameter to the job tempate, and creates another pair of
bonnyci-run-*-multinode jobs, so that we may continue using the
multinode label for our hoist checks.

This .bonnyci layout for the repo itself is driven by run-parts and
will execute any executable files in tests/, named ^test-.$ and in
lexical order.

A noop script and some symlinks have been added to tests/ so that
existing job definitions can pass and this can land. These
can be dropped once this patch lands.
