+++ test-requirements.txt
@@ -1,0 +1,2 @@
+ansible
+jenkins-job-builder
