+++ tests/noop.py
@@ -1,0 +1,4 @@
+#!/usr/bin/python
+# A shim to get this layout migration through the gate.
+# This can be removed in a follow-up patch.
+print 'noop'
